scheduling with rhf
